Generate nonces with cryptographically secure PRNG
The str_shuffle() uses the Mersenne Twister (as of PHP 7.1) random
number generator, which is predictable.

Since the whole point of nonces is to be a value that is not predictable
(and only to be used once), Mersenne Twister is unsuitable for this
task.

We now generate the nonce using random_int() and only emit characters of
the alphabet we previously had in $permitted_chars. Since I'm not the
author of the original code, I can only assume that the author wanted to
use this as the alphabet even though the value is run through
base64_encode() again.

I also renamed $permitted_chars to $permittedChars for consistency,
since most other variables use camelCase.

Signed-off-by: aszlig <aszlig@nix.build>